Showing how many times finalize() will run.
Develop a java class that has finalize method which displays “Finalize method called”. Create
another class which creates objects of the previous class and it uses the same object reference for
creating these objects. For example, if A1 is the class name, then the objects are created as
below:
A1 a = new A1();
a = new A1();
Topic
No. Topic Sub Topics
Min No.of
Assignments to be
Done
1 Java.lang
package
1. Assertions
2. Enums
3. Object Class, Behaviour of equals/hashcode
4. Variable Argument Lists
2
2 Java.util
package
1. Wrapper Classes, Auto boxing, Unboxing
2. Dates, Numbers, Currency 2
3 Java.io package
1. File I/O
2. Java Logging Framework
3. Serialization
2
4 Lambda
Expressions
1. Lambda Expressions
2. Streams API (Aggregate Operations )
2
Total Min No.of Assignments to be Done
8
Sensitivity: Internal & Restricted
a = new A1();
When the statement Runtime.getRuntime().gc() is invoked, how many times the finalize method
is called?